Minor clean-up
VERSION was previously used to easily identify the git commit of a running server. It is no longer being used. Other files related to this file:
- .githoos/pre-commit
- Makefile
- urls.py
- common.py

Also no longer redirecting stderr and stdout from gunicorn and qcluster so that these logs are captured in the docker logs.

Other unused files were also removed.