fix: getting locations within boundaries
instead of using contained which uses the bounding box of a geometry we
use within.
this was happening on YE that points which were not in the selected
province still end up being shown because they were inside the bounding
box of the province.
using within makes sure that the points need to be within the geometry