fix: add migration that were missing before
the order did update properly in the database since they all had the
same value.
the field that was introduced in this migration:
0038_auto_20201117_1559.py
did not update all the exisiting fields. so all had 0 and adding a
migration was recommended by the library here:
https://django-admin-sortable2.readthedocs.io/en/latest/usage.html#initial-data
the previously added tests shows that the ordering works as expected